feat: 상태 라이브러리 추가
